+++ custom_addons/ak_library_management/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     'name': 'Library Management',
-    'version': '18.0.1.0.1',
+    'version': '18.0.1.0.2',
     'author': 'Jay Butani',
     'sequence': 5,
     'summary': 'Library Management Module',
@@ -22,8 +22,8 @@
         'views/sale_menus.xml',
         'views/library_bulk_book_views.xml',
         'views/res_users_views.xml',
-        'wizard/sale_order_wizard_views.xml',
-        'views/sale_order_views.xml',
+        'views/sale_order_add_button_views.xml',
+        'wizard/sale_order_check_quantity_wizard_views.xml',
         'views/res_partner_views.xml',
         'views/borrow_transaction_history_views.xml',
         'wizard/borrow_transaction_history_wizard_views.xml',