+++ custom_addons/ak_library_management/wizard/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import borrow_transaction_history_wizard
-from . import sale_order_wizard
+from . import sale_order_check_quantity_wizard