+++ custom_addons/ak_library_management/wizard/sale_order_check_quantity_wizard.py
@@ -8,6 +8,6 @@ class SaleOrderWizard(models.TransientModel):
     create transient model and add message field to show the wizard message
     """
     _name = 'sale.order.wizard'
-    _description = 'sale order wizard'
+    _description = 'Sale order wizard'
 
     message = fields.Char(readonly=True)