+++ custom_addons/ak_library_management/changelog.md
@@ -1,7 +1,14 @@
-## [18.0.1.0.1] - 2025-03-06
+## [18.0.1.0.1] - 2025-03-06 | constraints mail & activity
 
 - Add proper docstring and create wizard directory
 - changes on mail template
 - Add mail composer on borrow transaction history
 - optimize the domain
 - Add field in user model
+
+## [18.0.1.0.2] - 2025-03-07 | method overriding
+
+- change in field name and rename some files
+- changes in condition of sale order action confirm
+- change in search method of borrow transaction model
+- delete automated action(duplicate name) from product template